null

+++ data_processing.py
@@ -41,7 +41,7 @@ def vertical_profile_plot(
     occultation_type_value = np.where((sunset_sunrise == 0) | (sunset_sunrise == 2), "Sunset", "Sunrise")
 
     # If statement to handle single date_index input
-    if isinstance(date_index, np.int64):
+    if isinstance(date_index, np.int32):
         current_date_index = date_index
         current_occultation = occultation_type_value[current_date_index]
